% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragility.R
\name{plot.fragility_curve}
\alias{plot.fragility_curve}
\title{Plot method for fragility_curve objects}
\usage{
\method{plot}{fragility_curve}(x, ...)
}
\arguments{
\item{x}{A \code{fragility_curve} data frame (output of
\code{\link{fragility_curve}}).}

\item{...}{Ignored.}
}
\value{
A combined \code{patchwork} plot if the \pkg{patchwork} package is
  installed, otherwise the two panels are printed separately and a list of
  two \code{ggplot2} objects is returned invisibly.
}
\description{
Produces a dual-panel plot: the dose-response curve (top) and the
fragility curve (bottom), with high-fragility regions shaded. If
\code{support_density} is present in \code{x} (i.e. \code{t_obs} was
supplied to \code{\link{fragility_curve}}), a scaled density ribbon is
overlaid on the fragility panel to flag low-support regions.
}
\examples{
dat <- simulate_dose_response(200, dgp = "threshold")
fit <- causal_spline(Y ~ T | X1 + X2 + X3, data = dat)
fc  <- fragility_curve(fit, t_obs = dat$T)
plot(fc)

}
