% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.causal_spline}
\alias{print.causal_spline}
\title{Print method for causal_spline objects}
\usage{
\method{print}{causal_spline}(x, ...)
}
\arguments{
\item{x}{A \code{causal_spline} object.}

\item{...}{Ignored.}
}
\description{
Print method for causal_spline objects
}
