% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.causal_spline}
\alias{summary.causal_spline}
\title{Summary method for causal_spline objects}
\usage{
\method{summary}{causal_spline}(object, ...)
}
\arguments{
\item{object}{A \code{causal_spline} object.}

\item{...}{Ignored.}
}
\description{
Summary method for causal_spline objects
}
